assignment2-in practice/requirements.txt
flask
flask-cors
PyJWT
redis

